New feature: sending animated emotes
top-level function was added to index.js, which allows a user to send a message with an animated emote through the bot by beginning the message with a dash "-".

Noted bugs: the bot will replace your message if it begins with a dash "-", even if there are no valid emotes. This command also does not work for multi-line messages.