bug fixes in index.js
Bot no longer replaces message if begun with a dash "-". Message is replaced if-and-only-if a valid animated emote is found and replaced into the string.

To help prevent accidental inputs, animated emote name must now be enclosed in backticks (e.g. `name`). This works even for emotes typed consequently (i.e. without spaces), irrespective of the resulting markdown text.

Event now functions for back-to-back emotes (i.e. without spaces), emotes in a message, and multi-line messages.

Issues: the new message is still sent by the bot and not a webhook with the user's appearance.