bug fix in index.js
Fixed bug where message is replaced if a message contains only the name of a valid animated emote, even if not enclosed by backticks. Backticks increase the array size and can be used as a check.

Function now returns the value of the new string, to be sent through a webhook.