new feature: emote lists + bugs found
The command !emotelist now sends an ordered list of static & animated emotes in a server, including their capacities.

This command, however, is currently restricted to my ID only, as the bot was added to other servers with no restrictions in place to prevent spamming.

Tried adding an option to send the list of emotes for a different server the bot is in, but difficult to implement.

Some bugs found in the emote command: does not always delete the command message (maybe permissions are missing); sometimes will delete the command message even if a valid emote is not found, and does not work for external emojis.