bug fix: emote command, duplicate name
When duplicate emote names are encountered, the bot will prompt the user to choose which version to use. It will get messy, quickly, if multiple people user this command.

Consider defaulting to the first match found, or allowing the user to select which to use preemptively (e.g. using `rinsway_2` to get the second match). This fails if an emote, for whatever reason, is naturally named this way.

Noted bugs: if there are multiple duplicate encounters (e.g. `rinsway` + `rinhair`, the user will answer for BOTH inputs, despite there being two separate prompts.

This command, for some reason, does not allow external emotes in Rin Simp Discord server, despite it functionality as normal in all other servers.