new message.js module
Functions for the emote command are now contained in the message.js file, which is required by index.js

Changes: command no longer prompts the user for an input, since it can make chat messy, especially with large number of uses; if multiple name matches are found, it will prioritize the emote in the current server and then the first match.

Potential issues: if the bot is in a large number of servers with overlapping emotes, the wrong emote may be prioritized; if a static and animated emote have the same name, there is no way to manually choose between them.