Folder restructuring + emote command update
Created the error, message, and ready events that are called into index. client.on events are properly exported from these event files.

Update to emote command, which now operates using the normal :emote: syntax. It will only attempt to replace the emote for non-nitro user, as nitro users will automatically convert the string into emote form.

Noted issues: the bot replaces the message even if the emote is contained within code (backticks). Need a check to prevent this.

Still issues with webhook in Rin Simp server. Appears to be server-side.