emote command, bug fixes
Added permission checks for CinnaBot and @everyone roles in the messaged server. Informs the user of the missing permissions, if any; otherwise, it deletes the original message and send a webhook as normal.

To check the client permissions, the function sendMessageWebhook has been moved into the module.exports.

By request, the emote name check is no longer case-sensitive; nitro users can now use the function being a mod.